create TimeCache (low-level functions from VMSCache)
