+++ VMSCache.py
@@ -19,10 +19,13 @@
 # You should have received a copy of the GNU General Public License
 # along with this program. If not, see <https://www.gnu.org/licenses/>.
 
+__all__ = ["VMSCache"]
+
+from TimeCache import *
 import numpy as np
 
 
-class VMSCache:
+class VMSCache(TimeCache):
     """Cache for large float data. Holds rolling time window of records. Act as
     dictionary, where keys are accelerometer number and axis
     (1X,1Y,1Z,..,<sensors>Z). [] and len operators are supported.
@@ -38,65 +41,23 @@ class VMSCache:
     """
 
     def __init__(self, size, sensors, window=3):
-        self._size = size
-        self._window = window
         self._sensors = sensors
         items = [("timestamp", "f8")] + [
             (f"{s} {a}", "f8")
             for s in range(1, self._sensors + 1)
             for a in ["X", "Y", "Z"]
         ]
-        self.data = np.zeros((self._size), items, order="F")
-        self.clear()
+
+        super().__init__(size, items, window)
+
+    def clear(self):
+        """Clear cache."""
+        super().clear()
+        self._receiving = []
 
     def sensors(self):
         """Returns number of sensors stored in cache."""
         return self._sensors
-
-    def clear(self):
-        """Clear cache."""
-        self.current_index = 0
-        self.filled = False
-        self._receiving = []
-
-    def resize(self, size):
-        """Change cache size. Data are preserved - either all rows are used
-        when expanding, or the most recent ones are stored when shrinking.
-
-        Parameters
-        ----------
-        size : `int`
-            New size.
-        """
-        clength = len(self)
-        if size == clength:
-            return
-        newdata = np.zeros(size, self.data.dtype)
-        n_current = r = min(clength, size)
-        for s in self.rows_reverse():
-            r -= 1
-            newdata[r] = s
-            if r == 0:
-                break
-        self.filled = (n_current == size) and (clength > 0)
-        self.current_index = n_current
-
-        self.data = newdata
-        self._size = size
-
-    def append(self, data):
-        """Append new row to end of data.
-
-        Parameters
-        ----------
-        data : `tupple`
-            New row data.
-        """
-        if self.current_index >= self._size:
-            self.current_index = 0
-            self.filled = True
-        self.data[self.current_index] = data
-        self.current_index += 1
 
     def newChunk(self, data, sample_period):
         """Add new data chunk. Append data to cache if all sensors are
@@ -172,51 +133,3 @@ class VMSCache:
             self._receiving = self._receiving[1:]
             chunk_removed = True
         return (added, chunk_removed)
-
-    def startTime(self):
-        """Return timestamp of the last data point.
-
-        Returns
-        -------
-        endTime : `float`
-            None if cache is empty. Otherwise timestamp of the first data point."""
-        if self.filled is False:
-            if self.current_index > 0:
-                return self.data[0]["timestamp"]
-            return None
-
-        if self.current_index >= self._size:
-            return self.data[0]["timestamp"]
-        return self.data[self.current_index]["timestamp"]
-
-    def endTime(self):
-        """Return timestamp of the last data point.
-
-        Returns
-        -------
-        endTime : `float`
-            None if cache is empty. Otherwise timestamp of the last data point."""
-        if self.current_index == 0:
-            if self.filled is False:
-                return None
-            return self.data[-1]["timestamp"]
-        return self.data[self.current_index - 1]["timestamp"]
-
-    def rows_reverse(self):
-        """Yelds reversed row iterator."""
-        for r in range(self.current_index - 1, -1, -1):
-            yield self.data[r]
-        if self.filled:
-            for r in range(self._size - 1, self.current_index - 1, -1):
-                yield self.data[r]
-
-    def __getitem__(self, key):
-        if self.filled:
-            return list(self.data[self.current_index + 1 :][key]) + list(
-                self.data[: self.current_index][key]
-            )
-        else:
-            return list(self.data[: self.current_index][key])
-
-    def __len__(self):
-        return self._size if self.filled else self.current_index
